Renaming vfam test files
