+++ tests/vfam/test_markov.py
@@ -3,7 +3,7 @@ import filecmp
 import os
 from pathlib import Path
 
-from fixtures_vfam import *
+from fixtures import *
 from paths_vfam import VFAM_INTERMEDIATES_PATH, DUPES_INPUT, GENERIC_INPUT, LARGE_INPUT
 
 BLAST_DUPES = VFAM_INTERMEDIATES_PATH / "Dupes" / "blast"
