+++ tests/vfam/test_polyprotein.py
@@ -1,7 +1,7 @@
 import pytest
 from pathlib import Path
 
-from fixtures_vfam import *
+from fixtures import *
 from paths_vfam import VFAM_INTERMEDIATES_PATH, DUPES_INPUT, GENERIC_INPUT, LARGE_INPUT
 
 POLYP_DUPES = VFAM_INTERMEDIATES_PATH / "Dupes" / "polyprotein_ids"
